refactor(ui): extract icon and user-avatar from its directories
