--- app/src/components/ui/icon/index.ts
@@ -1,1 +1,0 @@
-export { Icon } from './icon';
