--- app/src/components/ui/user-avatar/index.ts
@@ -1,1 +1,0 @@
-export { UserAvatar } from './user-avatar';
